Update repo script
